Add support for file: URIs
This is bare support for #23, but it's not optimal. Use with
care.

+++ app/src/main/java/org/jak_linux/dns66/MainActivity.java
@@ -187,7 +187,7 @@ public class MainActivity extends AppCompatActivity {
         for (Configuration.Item item : config.hosts.items) {
             File file = FileHelper.getItemFile(this, item);
 
-            if (file != null && item.state != 2) {
+            if (item.isDownloadable() && file != null && item.state != 2) {
                 DownloadManager.Request request = new DownloadManager.Request(Uri.parse(item.location));
                 Log.d("MainActivity", String.format("refresh: Downkoading %s to %s", item.location, file.getAbsolutePath()));
                 file.delete();
